adding mail notification after user approval, ban, unban receptionist and approve user

+++ app/Http/Controllers/ManagersController.php
@@ -11,39 +11,35 @@ use Illuminate\Database\Capsule\Manager;
 use Illuminate\Support\Facades\Validator;
 
 
-
 class ManagersController extends Controller
 {
     public function index(ManagersDatatable $dataTable)
-    { 
-        $managers = User::with(['roles' => function($q){
-        $q->where('name', 'receptionist');
-         }])->get();
-        // return Datatables::of($receptionists); 
-        return $dataTable->render('admin.managers.index',['title' => 'Managers', 'managers' => $managers]);
+    {
+        return $dataTable->render('admin.managers.index', ['title' => 'Managers']);
     }
 
-    public function create(){
+    public function create()
+    {
 
     }
 
     public function store(Request $request)
-    {   
+    {
         $rules = [
             'name' => 'required',
             'email' => 'required|unique:users',
             'national_id' => 'required|unique:users',
-            'avatar'=>'image|mimes:jpeg,jpg|max:1999',
+            'avatar' => 'image|mimes:jpeg,jpg|max:1999',
             'mobile' => 'required|unique:users',
             'country' => 'required',
             'password' => 'min:6',
         ];
 
-        if($request->hasFile('avatar')){
-            $file=$request->file('avatar');
-            $ext=$file->getClientOriginalExtension();
-            $filename="img" . "_" . time() . "." . $ext;
-            $file->storeAs('public/images',$filename);
+        if ($request->hasFile('avatar')) {
+            $file = $request->file('avatar');
+            $ext = $file->getClientOriginalExtension();
+            $filename = "imges" . "_" . time() . "." . $ext;
+            $file->storeAs('public/images', $filename);
         }
         $validator = Validator::make($request->all(), $rules);
         if ($validator->fails()) {
@@ -54,25 +50,26 @@ class ManagersController extends Controller
         }
 
 
-        $user=User::create([
-            'name'  => $request->name,
-            'email'  => $request->email,
+        $user = User::create([
+            'name' => $request->name,
+            'email' => $request->email,
             'national_id' => $request->national_id,
             'password' => $request->password,
-            'avatar'  => $filename,
-            'country'  => $request->country,
-            'mobile'  => $request->mobile,
-            'gender'  => $request->gender,
-            'approved'=>True,
-            'approved_by'=> auth()->user()->id
+            'avatar' => $filename,
+            'country' => $request->country,
+            'mobile' => $request->mobile,
+            'gender' => $request->gender,
+            'approved' => true,
+            'approved_by' => auth()->user()->id
         ]);
-       if($user->exists()){
-         $user->assignRole('manager');  
-         return response()->json(array('success' => true), 200);
-       }
-       return  response()->json(array('success' => false), 400);
+        if ($user->exists()) {
+            $user->assignRole('manager');
+            return response()->json(array('success' => true), 200);
+        }
+        return response()->json(array('success' => false), 400);
 
     }
+
     public function edit($id)
     {
         if (\request()->ajax()) {
@@ -81,24 +78,25 @@ class ManagersController extends Controller
                 return \response()->json($manager);
         }
     }
+
     public function update(Request $request, $id)
     {
         // validation
         $rules = [
             'name' => 'required',
-            'email'     => 'required|email|unique:users,email,' .$id,
-            'national_id' => 'required|digits_between:10,17|unique:users,national_id,' .$id,
-            'avatar'=>'image|mimes:jpeg,jpg|max:1999',
-            'mobile' => 'required|regex:/(01)[0-9]{9}/|unique:users,mobile,' .$id,
+            'email' => 'required|email|unique:users,email,' . $id,
+            'national_id' => 'required|digits_between:10,17|unique:users,national_id,' . $id,
+            'avatar' => 'image|mimes:jpeg,jpg|max:1999',
+            'mobile' => 'required|regex:/(01)[0-9]{9}/|unique:users,mobile,' . $id,
             'country' => 'required',
             'password' => 'min:6',
-          
+
         ];
 
         $validator = Validator::make($request->all(), $rules);
-       
+
         if ($validator->fails()) {
-        
+
             return \response()->json([
                 'success' => false,
                 'errors' => $validator->getMessageBag()->toArray()
@@ -106,12 +104,12 @@ class ManagersController extends Controller
         }
 
         $manager = User::find($id);
-       
 
-         $manager->update($request->all());
+
+        $manager->update($request->all());
         return \response()->json(array('success' => true), 200);
     }
-   
+
     /**
      * Remove the specified resource from storage.
      *
@@ -120,8 +118,9 @@ class ManagersController extends Controller
      */
 
     public function destroy($id)
-    {    if (request()->ajax()) {
-           $user=User::find($id);
+    {
+        if (request()->ajax()) {
+            $user = User::find($id);
             if ($user->delete()) {
                 return response('success');
             }
